Created an abstraction over containerization platforms (like docker or singularity/apptainer), as well as an abstraction over workflow engines like REANA, etc... Since REANA most likely doesnt support GPU access, nor does it support any other HPC besides the CERN one, we may need to have a different workflow engines and containerization platforms for each backend architecture...
