+++ backend/app/main.py
@@ -12,7 +12,11 @@ from app.models.experiment import Experiment
 from app.models.experiment_run import ExperimentRun
 from app.models.experiment_template import ExperimentTemplate
 from app.routers import aiod, experiment_templates, experiments
-from app.services.experiment import ExperimentService
+from app.services.container_platforms.base_platform import ContainerBasePlatform
+from app.services.container_platforms.docker import DockerService
+from app.services.scheduling import ExperimentScheduling
+from app.services.workflow_engines.base_engine import WorkflowBaseEngine
+from app.services.workflow_engines.reana import ReanaService
 
 app = FastAPI(title="AIOD - Practitioner's Portal", version=__version__)
 
@@ -39,20 +43,23 @@ async def app_init():
     app.db = AsyncIOMotorClient(settings.MONGODB_URI, uuidRepresentation="standard")[
         settings.MONGODB_DBNAME
     ]
-
     await init_beanie(
         database=app.db,
         document_models=[ExperimentTemplate, Experiment, ExperimentRun],
     )
 
-    # Setup DockerClient and create queue of ExperimentRuns to execute
-    docker_service = await ExperimentService.init_docker_service()
+    # initialize container platform and worfklow engine
+    container_platform: ContainerBasePlatform = await DockerService.init()
+    workflow_engine: WorkflowBaseEngine = await ReanaService.init()
 
-    # Create task for scheduling docker image builds for ExperimentTemplates from queue
-    asyncio.create_task(docker_service.schedule_image_building())
+    # Setup ExperimentScheduling and create queues of experiments and images to execute
+    experiment_scheduling = await ExperimentScheduling.init(
+        container_platform, workflow_engine
+    )
 
-    # Create task for scheduling ExperimentRuns from queue
-    asyncio.create_task(docker_service.schedule_experiment_runs())
+    # Create seperate tasks for scheduling experiments and images
+    asyncio.create_task(experiment_scheduling.schedule_image_building())
+    asyncio.create_task(experiment_scheduling.schedule_experiment_runs())
 
 
 @app.on_event("shutdown")
@@ -60,7 +67,7 @@ async def shutdown_event():
     if getattr(app, "db", None) is not None:
         app.db.client.close()
 
-    ExperimentService.get_docker_service().close_docker_connection()
+    await DockerService.get_service().terminate()
 
     await aiod_client_wrapper.stop()
